fire update... fixed output to /dev/kmsg and also added in the filtering of 'mathuez' for example(name of module) in /proc/kallsyms to hide all hooks
